Back to .cm but clean up namespace a bit
