+++ cmcrameri/__init__.py
@@ -5,11 +5,13 @@ http://www.fabiocrameri.ch/colourmaps.php
 See README.md for an overview and instructions
 """
 from __future__ import absolute_import
-from ._cm import cm, paths, show_cmaps
+
+from . import cm
+from .cm import show_cmaps
+
 
 __all__ = (
     "cm",
-    "paths",
     "show_cmaps",
 )
